+++ dags/atd_knack_signal_work_orders.py
@@ -1,12 +1,26 @@
 from os import getenv
 
-from airflow.models import DAG
-from airflow.operators.docker_operator import DockerOperator
+from airflow.sdk import DAG
+from airflow.providers.docker.operators.docker import DockerOperator
 from pendulum import datetime, duration
 
 from utils.onepassword import get_env_vars_task
 from utils.knack import get_date_filter_arg
 from utils.slack_operator import task_fail_slack_alert
+
+doc_md = """
+## Knack Services DAG: Traffic Signal Work Orders
+
+Load signal work orders (view_1829) records from Knack to Postgrest to Socrata
+
+## Troubleshooting
+
+Most of the time just re-triggering this DAG will likely resolve any issues automatically.
+
+The most common bug is when the underlying Knack view is changed and the corresponding Socrata dataset
+was not updated to match.
+
+"""
 
 DEPLOYMENT_ENVIRONMENT = getenv("ENVIRONMENT", "development")
 
@@ -17,26 +31,18 @@ DEFAULT_ARGS = {
     "email_on_failure": False,
     "email_on_retry": False,
     "retries": 0,
-    "execution_timeout": duration(minutes=30),
+    "execution_timeout": duration(minutes=5),
     "on_failure_callback": task_fail_slack_alert,
 }
 
 REQUIRED_SECRETS = {
     "KNACK_APP_ID": {
-        "opitem": "Knack Signs and Markings",
-        "opfield": f"production.appId",
+        "opitem": "Knack AMD Data Tracker",
+        "opfield": "production.appId",
     },
     "KNACK_API_KEY": {
-        "opitem": "Knack Signs and Markings",
-        "opfield": f"production.apiKey",
-    },
-    "PGREST_ENDPOINT": {
-        "opitem": "atd-knack-services PostgREST",
-        "opfield": "production.endpoint",
-    },
-    "PGREST_JWT": {
-        "opitem": "atd-knack-services PostgREST",
-        "opfield": "production.jwt",
+        "opitem": "Knack AMD Data Tracker",
+        "opfield": "production.apiKey",
     },
     "SOCRATA_API_KEY_ID": {
         "opitem": "Socrata Key ID, Secret, and Token",
@@ -50,27 +56,36 @@ REQUIRED_SECRETS = {
         "opitem": "Socrata Key ID, Secret, and Token",
         "opfield": "socrata.appToken",
     },
+    "PGREST_ENDPOINT": {
+        "opitem": "atd-knack-services PostgREST",
+        "opfield": "production.endpoint",
+    },
+    "PGREST_JWT": {
+        "opitem": "atd-knack-services PostgREST",
+        "opfield": "production.jwt",
+    },
 }
 
 
 with DAG(
-    dag_id="atd_knack_signs_markings_311_csr",
-    description="Publish signs markings 311 CSRs to Postgres, Socrata",
+    dag_id="atd_knack_signal_work_orders",
+    description="Load signal work orders (view_1829) records from Knack to Postgrest to Socrata",
+    doc_md=doc_md,
     default_args=DEFAULT_ARGS,
-    schedule_interval="20 19 * * *" if DEPLOYMENT_ENVIRONMENT == "production" else None,
-    tags=["repo:atd-knack-services", "knack", "socrata", "signs-markings"],
+    schedule="50 8 * * *" if DEPLOYMENT_ENVIRONMENT == "production" else None,
+    tags=["repo:atd-knack-services", "knack", "socrata", "data-tracker"],
     catchup=False,
 ) as dag:
     docker_image = "atddocker/atd-knack-services:production"
-    app_name = "signs-markings"
-    container = "view_3992"
+    app_name = "data-tracker"
+    container = "view_1829"
 
-    date_filter_arg = get_date_filter_arg(should_replace_monthly=True)
+    date_filter_arg = get_date_filter_arg(should_replace_monthly=False)
 
     env_vars = get_env_vars_task(REQUIRED_SECRETS)
 
     t1 = DockerOperator(
-        task_id="smd_311_csrs_to_postgrest",
+        task_id="atd_knack_signal_work_orders_to_postgrest",
         image=docker_image,
         docker_conn_id="docker_default",
         auto_remove="force",
@@ -82,7 +97,7 @@ with DAG(
     )
 
     t2 = DockerOperator(
-        task_id="smd_311_csrs_to_socrata",
+        task_id="atd_knack_signal_work_orders_to_socrata",
         image=docker_image,
         docker_conn_id="docker_default",
         auto_remove="force",
