+++ dags/atd_knack_traffic_detectors_weekly_snapshot.py
@@ -1,12 +1,25 @@
 from os import getenv
 
-from airflow.models import DAG
-from airflow.operators.docker_operator import DockerOperator
+from airflow.sdk import DAG
+from airflow.providers.docker.operators.docker import DockerOperator
 from pendulum import datetime, duration
 
 from utils.onepassword import get_env_vars_task
-from utils.knack import get_date_filter_arg
 from utils.slack_operator import task_fail_slack_alert
+
+doc_md = """
+## Knack Services DAG: Traffic Detectors Weekly Snapshot
+
+Appends traffic detector assets to a running log in Socrata
+
+## Troubleshooting
+
+Most of the time just re-triggering this DAG will likely resolve any issues automatically.
+
+The most common bug is when the underlying Knack view is changed and the corresponding Socrata dataset
+was not updated to match.
+
+"""
 
 DEPLOYMENT_ENVIRONMENT = getenv("ENVIRONMENT", "development")
 
@@ -16,19 +29,20 @@ DEFAULT_ARGS = {
     "start_date": datetime(2015, 1, 1, tz="America/Chicago"),
     "email_on_failure": False,
     "email_on_retry": False,
-    "retries": 0,
-    "execution_timeout": duration(minutes=30),
+    "retries": 5,
+    "retry_delay": duration(minutes=5),
+    "execution_timeout": duration(minutes=10),
     "on_failure_callback": task_fail_slack_alert,
 }
 
 REQUIRED_SECRETS = {
     "KNACK_APP_ID": {
         "opitem": "Knack AMD Data Tracker",
-        "opfield": f"production.appId",
+        "opfield": "production.appId",
     },
     "KNACK_API_KEY": {
         "opitem": "Knack AMD Data Tracker",
-        "opfield": f"production.apiKey",
+        "opfield": "production.apiKey",
     },
     "SOCRATA_API_KEY_ID": {
         "opitem": "Socrata Key ID, Secret, and Token",
@@ -50,27 +64,41 @@ REQUIRED_SECRETS = {
         "opitem": "atd-knack-services PostgREST",
         "opfield": "production.jwt",
     },
+    "AWS_ACCESS_ID": {
+        "opitem": "Socrata Dataset Backups S3 Bucket",
+        "opfield": "production.AWS Access Key",
+    },
+    "AWS_SECRET_ACCESS_KEY": {
+        "opitem": "Socrata Dataset Backups S3 Bucket",
+        "opfield": "production.AWS Secret Access Key",
+    },
+    "BUCKET": {
+        "opitem": "Socrata Dataset Backups S3 Bucket",
+        "opfield": "production.Bucket",
+    },
 }
 
 
 with DAG(
-    dag_id="atd_knack_signal_detection_status_log",
-    description="Load signal detection status log (view_3003) records from Knack to Postgrest to Socrata",
+    dag_id="atd_knack_traffic_detectors_weekly_snapshot",
+    description="Appends traffic detector assets to a running log in Socrata",
+    doc_md=doc_md,
     default_args=DEFAULT_ARGS,
-    schedule_interval="0 5 * * *" if DEPLOYMENT_ENVIRONMENT == "production" else None,
-    tags=["repo:atd-knack-services", "knack", "socrata", "data-tracker"],
+    schedule="0 1 * * MON" if DEPLOYMENT_ENVIRONMENT == "production" else None,
+    tags=["repo:atd-knack-services", "knack", "socrata"],
     catchup=False,
 ) as dag:
     docker_image = "atddocker/atd-knack-services:production"
     app_name = "data-tracker"
-    container = "view_3003"
+    container = "view_4274"
 
-    date_filter_arg = get_date_filter_arg(should_replace_monthly=True)
+    # we always want to append the the complete view contents every time
+    date_filter_arg = "-d 1970-01-01"
 
     env_vars = get_env_vars_task(REQUIRED_SECRETS)
 
     t1 = DockerOperator(
-        task_id="atd_knack_traffic_signal_detection_status_log_to_postgrest",
+        task_id="atd_knack_traffic_detectors_weekly_snapshot_to_postgrest",
         image=docker_image,
         docker_conn_id="docker_default",
         auto_remove="force",
@@ -82,7 +110,7 @@ with DAG(
     )
 
     t2 = DockerOperator(
-        task_id="atd_knack_traffic_signal_detection_status_log_to_socrata",
+        task_id="atd_knack_traffic_detectors_weekly_snapshot_to_socrata",
         image=docker_image,
         docker_conn_id="docker_default",
         auto_remove="force",
@@ -92,4 +120,15 @@ with DAG(
         mount_tmp_dir=False,
     )
 
-    date_filter_arg >> t1 >> t2
+    t3 = DockerOperator(
+        task_id="atd_knack_traffic_detectors_weekly_snapshot_socrata_backup",
+        image=docker_image,
+        docker_conn_id="docker_default",
+        auto_remove="force",
+        command=f"./atd-knack-services/services/backup_socrata.py -a {app_name} -c {container}",
+        environment=env_vars,
+        tty=True,
+        mount_tmp_dir=False,
+    )
+
+    t1 >> t2 >> t3
